requirements.txt
Flask==2.3.2
pymongo==4.6.1
requests==2.31.0
python-dotenv==1.0.1
matplotlib==3.7.1
xhtml2pdf==0.2.12
numpy==1.24.4
reportlab==3.6.12